requirements.txt
anaconda-client==1.6.3
appnope==0.1.0
beautifulsoup4==4.6.0
bleach==1.5.0
certifi==2016.2.28
click==6.7
clyent==1.2.2
decorator==4.1.2
docopt==0.6.2
entrypoints==0.2.3
Flask==0.12.2
grip==4.4.0
html5lib==0.999
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
Markdown==2.6.9
MarkupSafe==1.0
mistune==0.7.4
nb-anacondacloud==1.4.0
nb-conda==2.2.0
nb-conda-kernels==2.1.0
nbconvert==5.2.1
nbformat==4.4.0
nbpresent==3.0.2
notebook==5.0.0
numpy==1.13.1
olefile==0.44
pandas==0.20.3
pandocfilters==1.4.2
path-and-address==2.0.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
requests==2.14.2
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3
tornado==4.5.2
traitlets==4.3.2
Unidecode==0.4.21
wcwidth==0.1.7
Werkzeug==0.12.2
widgetsnbextension==3.0.2
